requirements.txt
# --- Базовые зависимости проекта ---
langchain>=0.2.0
langchain-chroma>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.2.0
langchain-huggingface>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.22
openai>=1.3.5
tiktoken>=0.5.2
langfuse

# --- Модели эмбеддингов ---
sentence-transformers>=2.2.2
transformers>=4.38.0
huggingface-hub>=0.20.3

# --- Парсинг данных ---
beautifulsoup4>=4.12.2
requests>=2.31.0
playwright>=1.40.0

# --- Обработка изображений ---
easyocr>=1.7.1
pillow>=10.0.1
torch>=2.1.0

# --- OCR + BLIP ---
opencv-python-headless>=4.8.1.78

# --- Утилиты ---
python-dotenv>=1.0.0
pathlib>=1.0.1